`RAIIMix_Chunk`: value-initialize base `Mix_Chunk`
Currently, `RAIIMix_Chunk` is only used at global scope, so the loader
ensures that the members are all zero.  `RAIIMix_Chunk::~RAIIMix_Chunk`
assumes that `abuf != nullptr` means that `abuf` must be freed.  Change
the constructor to value-initialize the base type, so that a
`RAIIMix_Chunk` defined at local scope will set `abuf` to `nullptr` on
construction.  Only `abuf` needs to be cleared, but clearing the entire
structure is cheap in local scope, and free in global scope, since the
compiler recognizes that it can skip writing zeroes and rely on the
loader to zero the variable.

This change has no effect now, but protects against a local instance of
`RAIIMix_Chunk` calling `delete[]` on an uninitialized pointer.