Remove leading dot from object file names
Commit 3fe3da9a259d855de0065aaf298544853a2d42fd introduced this pattern,
back when the expectation was that most users would not set a build
directory, and so object files would be mixed in with source files.
Commit a7f2d0a7513c766d502fead4f8e2e2af87aba8b3 changed the default to
use a directory named `build` if the user did not set a build directory.
As such, there is no need to continue hiding object files.  Remove their
leading dot, to simplify using shell globs to match them.