`player_has_primary_weapon`: remove hack that handled negative player energy
The originally released Descent 1 source included a function, then named
`player_has_weapon`, which began:

```
	//	Hack! If energy goes negative, you can't fire a weapon that doesn't require energy.
	//	But energy should not go negative (but it does), so find out why it does!
	if (Players[Player_num].energy < 0)
		Players[Player_num].energy = 0;
```
The logic was changed in c483deb646950283359e9b6266cc117ea910ef77[1] to
leave the energy negative, but override the test and let the player fire
anyway.  In the 9 years since that commit, no one has ever reported
seeing player energy go negative.  Since the comment and associated
workaround date back to the Descent 1 source release, it is unclear
when this workaround was last needed, or if this was a remnant from a
bug fixed long ago.  Recent work reviewing updates to player energy
revealed one possible bug site related to triggers, which was fixed in
2bf6761e665b501c68b7d29cbf26fc360fed238d[2].  As such, this hack is
likely unnecessary now.  Remove it.

[1]: c483deb646950283359e9b6266cc117ea910ef77 ("Change Parallax negative energy hack to avoid modifying energy")
[2]: 2bf6761e665b501c68b7d29cbf26fc360fed238d ("`check_trigger_sub`: clamp D1 energy drain to avoid negative energy")