Tighten check for `inet_ntop`
On Windows, `inet_ntop` may be absent[1], so `SConstruct` tests for its
presence.  When it is present, it will need to be linked, so extend the
test to require that the test program link successfully, rather than
only testing compilation, as was done previously.

[1]: `inet_ntop` is mandatory for IPv6 support and is present as of
Windows Vista.  I think it is unlikely anyone still builds Rebirth for
Windows with a pre-Windows Vista target (_WIN32_WINNT < 0x600).
However, this support builds successfully, so there is no reason to
remove it yet.