Fix `inet_ntop` test on >=gcc-11
gcc-11 enhanced -Wmaybe-uninitialized to warn when passing the address
of an uninitialized object as `const T *`, on the assumption that the
called function will read from the pointer, and get uninitialized data
as a result.  This test program never runs, so no uninitialized access
ever occurs.  However, the warning about potential uninitialized usage
causes the test to report the wrong result.  Fix this by initializing
the `sockaddr_in` passed to `inet_ntop`.