Remove D1 use of `convert_tmap` on object load
D1X-Rebirth had the ability to coerce invalid textures to valid ones via
`convert_tmap`, starting in
commit da7fd57332bc8ca4ee3979af6c101a06f224982a ("merge loadrl2.c into
gamemine.c and gamesave.c").  This feature was not present in original
Descent 1.  D2X-Rebirth also lacked this feature.  Remove it, and
replace it with more general logic to map any invalid texture to the
no-override value, rather than picking a valid texture via modulo.
Since the feature was not in retail Descent 1, levels from that era
should not need it.  Maps made afterward ought to set their texture
overrides correctly, and not rely on the coercion.  Any levels which do
rely on coercion would be broken when loaded in D2X-Rebirth via its
emulation, since D2X-Rebirth does not perform coercion.