Reorder player-hit-wall damage computation
For D1X-Rebirth: no change.

For D2X-Rebirth: no observable change in gameplay.  Previously, the
damage would be computed, then the game would check to see if the wall
was a special wall (lava or water).  If it was, then the damage would be
reset to 0.  Reorder the logic to test for a special wall first, and
skip computing damage if the wall is special.