Validate some texture index values before use
Indexing beyond the end of Textures is undefined behavior.  Change some
uses to skip the read, and any steps that depended on the data loaded
from that read.  This should have no observable effect in game, since
any values that trigger the new skipping logic would previously have
been an out-of-bounds access, and would have crashed if array bounds
checking was used.