GitHub Actions: Windows: set RC=windres
Gentoo crossdev provides the Windows Resource Compiler as
`$CHOST-windres`.  The toolchain used by GitHub Actions provides it as a
bare `windres`.  Therefore, the addition of `CHOST=` caused `SConstruct`
to use a filename which does not exist.  Add an explicit override to
switch to the unqualified name used by GitHub Actions.